Some refactor after merge
