+++ ReactApp/src/models/ChartData.ts
